Refactor C++ gen to centralize rules for types
